Undoing fix for Indexing failure due to hashCode call on PdxInstance
Undoing the fix to call markIdentityField when serialization the session
object. What we discovered is that the geode index is trying to compute the
hashcode of the session object during an update to an existing geode entry.

Computing the hashCode computes the hashCode of *all* of the fields. In the
case of GemFireSession, some of those fields are not intended, and in fact
cannot be deserialized on the server.